+++ src/main/kotlin/net/wojteksz128/jsoupTest/model/Computer.kt
@@ -1,3 +1,3 @@
 package net.wojteksz128.jsoupTest.model
 
-data class Computer(val id: Int, val name: String, val url: String, val price: Int)
+data class Computer(val id: Int, val name: String, val url: String, val price: Double)
